chore: move useStateWithStorage hook into App.js
