+++ src/App.jsx
@@ -4,13 +4,13 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { AddItem, Home, Layout, List } from './views';
 
 import { getItemData, streamListItems } from './api';
-import { useStateWithStorage } from './utils';
 
 export function App() {
 	const [data, setData] = useState([]);
 	/**
 	 * Here, we're using a custom hook to create `listToken` and a function
-	 * that can be used to update `listToken` later.
+	 * that can be used to update `listToken` later. This hook ensures that our
+	 * application's state is synced with localStorage.
 	 *
 	 * `listToken` is `my test list` by default so you can see the list
 	 * of items that was prepopulated for this project.
@@ -58,4 +58,24 @@ export function App() {
 			</Routes>
 		</Router>
 	);
+}
+
+/**
+ * Set some state in React, and also persist that value in localStorage.
+ * @param {string} storageKey The key of the value in localStorage.
+ * @param {string | null} initialValue The initial value to store in localStorage and React state.
+ * @returns {[string | null, React.Dispatch<string | null>]}
+ */
+export function useStateWithStorage(storageKey, initialValue) {
+	const [value, setValue] = useState(
+		() => localStorage.getItem(storageKey) ?? initialValue,
+	);
+	useEffect(() => {
+		if (value === null || value === undefined) {
+			return localStorage.removeItem(storageKey);
+		}
+		return localStorage.setItem(storageKey, value);
+	}, [storageKey, value]);
+
+	return [value, setValue];
 }
